requirements.txt
numpy
scipy
opencv-python
scikit-image
matplotlib
Pillow
